Utils: migrate block explorer and token icon utilities to separate files
