+++ src/lib/README.md
@@ -1,0 +1,3 @@
+# `lib/`
+
+The `lib/` directory consists of internal utilities that are not exported to consumers of aragonUI.
