+++ src/utils/index.js
@@ -1,4 +1,5 @@
 export { default as color } from './color'
+export * from './block-explorer'
 export * from './components'
 export * from './contains-component'
 export * from './css'
@@ -9,5 +10,6 @@ export * from './format'
 export * from './keycodes'
 export * from './math'
 export * from './miscellaneous'
+export * from './token'
 export * from './url'
 export * from './web3'
