+++ src/utils/web3.js
@@ -1,99 +1,9 @@
 import sha3 from 'js-sha3'
-import { warn, warnOnce } from './environment'
 
 const { keccak_256: keccak256 } = sha3
 
-const EMPTY_ADDRESS = '0x0000000000000000000000000000000000000000'
 const TRANSACTION_REGEX = /^0x[A-Fa-f0-9]{64}$/
 const ADDRESS_REGEX = /^0x[0-9a-fA-F]{40}$/
-
-const TRUST_WALLET_BASE_URL =
-  'https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/'
-
-const TRUST_WALLET_NETWORKS = new Map([
-  // [chainId, path]
-  [1, 'ethereum'],
-  [100, 'xdai'],
-])
-
-const BLOCKSCOUT_NETWORKS = new Map([
-  // [chainId, path]
-  [6, 'etc/kotti'],
-  [61, 'etc/mainnet'],
-  [63, 'etc/mordor'],
-  [77, 'poa/sokol'],
-  [99, 'poa/core'],
-  [100, 'poa/xdai'],
-])
-const BLOCKSCOUT_TYPES = new Map([
-  ['block', 'blocks'],
-  ['transaction', 'tx'],
-  ['address', 'address'],
-  ['token', 'tokens'],
-])
-
-const ETHERSCAN_NETWORKS = new Map([
-  // [chainId, subdomain]
-  [1, ''],
-  [3, 'ropsten.'],
-  [4, 'rinkeby.'],
-  [42, 'kovan.'],
-  [5, 'goerli.'],
-])
-const ETHERSCAN_TYPES = new Map([
-  ['block', 'block'],
-  ['transaction', 'tx'],
-  ['address', 'address'],
-  ['token', 'token'],
-])
-
-const BLOCK_EXPLORERS = {
-  blockscout: {
-    name: 'Blockscout',
-    url: ({ chainId, type, value }) => {
-      if (!BLOCKSCOUT_NETWORKS.has(chainId)) {
-        throw new Error(`BlockscoutUrl chainId '${chainId}' not supported.`)
-      }
-      if (!BLOCKSCOUT_TYPES.has(type)) {
-        throw new Error(`BlockscoutUrl type '${type}' not supported.`)
-      }
-
-      const path = BLOCKSCOUT_NETWORKS.get(chainId)
-      const typePart = BLOCKSCOUT_TYPES.get(type)
-      return `https://blockscout.com/${path}/${typePart}/${value}`
-    },
-  },
-  etherscan: {
-    name: 'Etherscan',
-    url: ({ chainId, type, value }) => {
-      if (!ETHERSCAN_NETWORKS.has(chainId)) {
-        throw new Error(`EtherscanUrl chainId '${chainId}' not supported.`)
-      }
-      if (!ETHERSCAN_TYPES.has(type)) {
-        throw new Error(`EtherscanUrl type '${type}' not supported.`)
-      }
-
-      const subdomain = ETHERSCAN_NETWORKS.get(chainId)
-      const typePart = ETHERSCAN_TYPES.get(type)
-      return `https://${subdomain}etherscan.io/${typePart}/${value}`
-    },
-  },
-}
-
-// Shim mapping of web3.js' network type to chain ID.
-// This is mostly a convenience fallback for old users who may still be
-// specifying their desired network through a network type
-const NETWORK_TYPES_TO_CHAIN_ID = new Map([
-  ['main', 1],
-  ['ropsten', 3],
-  ['rinkeby', 4],
-  ['kovan', 42],
-  ['goerli', 5],
-
-  // Note that xDai is technically considered a "private" chain for web3.js,
-  // but we use "xdai" here to provide network detection
-  ['xdai', 100],
-])
 
 /**
  * Converts to a checksum address
@@ -107,7 +17,7 @@ const NETWORK_TYPES_TO_CHAIN_ID = new Map([
  * @param {String} address the given HEX address
  * @returns {String}
  */
-function toChecksumAddress(address) {
+export function toChecksumAddress(address) {
   if (!/^(0x)?[0-9a-f]{40}$/i.test(address)) {
     throw new Error(
       'Given address "' + address + '" is not a valid Ethereum address.'
@@ -198,96 +108,4 @@ export function isAddress(address) {
  */
 export function isTransaction(transaction) {
   return TRANSACTION_REGEX.test(transaction)
-}
-
-/**
- * Get the name of a block explorer
- *
- * @param {string} provider the explorer provider (e.g. etherscan).
- * @returns {string} The explorer's name, if any.
- */
-export function blockExplorerName(provider) {
-  const explorer = BLOCK_EXPLORERS[provider]
-
-  if (!explorer) {
-    warn(`blockExplorerUrl(): provider '${provider}' not supported.`)
-    return ''
-  }
-
-  return explorer.name
-}
-
-/**
- * Generates an etherscan URL
- *
- * @param {string} type The type of URL (block, transaction, address or token).
- * @param {string} value Identifier of the object, depending on the type (block number, transaction hash, …).
- * @param {object} options The optional parameters.
- * @param {string} options.chainId The EVM chain ID (https://chainid.network/).
- * @param {string} options.provider The explorer provider (e.g. etherscan).
- * @returns {string} The generated URL, or an empty string if the parameters are invalid.
- */
-export function blockExplorerUrl(type, value, options = {}) {
-  const { provider = 'etherscan' } = options
-
-  const explorer = BLOCK_EXPLORERS[provider]
-
-  if (!explorer) {
-    warn(`blockExplorerUrl(): provider '${provider}' not supported.`)
-    return ''
-  }
-
-  let { chainId = 1 } = options
-  if (!options.chainId && options.networkType) {
-    if (!NETWORK_TYPES_TO_CHAIN_ID.has(options.networkType)) {
-      warnOnce(
-        'blockExplorerUrl():validNetworkType',
-        'blockExplorerUrl() was used with a network type. It is recommended to use chainId instead.'
-      )
-      chainId = NETWORK_TYPES_TO_CHAIN_ID.get(options.networkType)
-    } else {
-      warn(
-        'blockExplorerUrl() was used with an invalid network type. It is recommended to use chainId instead.'
-      )
-
-      return ''
-    }
-  }
-
-  try {
-    return explorer.url({ chainId, type, value })
-  } catch (err) {
-    warn(`blockExplorerUrl(): ${err.message}`)
-    return ''
-  }
-}
-
-/**
- * Get the address of a token icon
- *
- * @param {string} address The contract address of the token, or the zero address (0x000…) to get the Ethereum icon.
- * @param {object} options The optional parameters.
- * @param {string} options.chainId The EVM chain ID (https://chainid.network/).
- * @returns {string} The generated URL, or an empty string if the parameters are invalid.
- */
-export function tokenIconUrl(address = '', { chainId = 1 } = {}) {
-  if (!TRUST_WALLET_NETWORKS.has(chainId)) {
-    warn(`tokenIconUrl(): invalid chainId '${chainId}'`)
-    return ''
-  }
-  const baseUrl = `${TRUST_WALLET_BASE_URL}/${TRUST_WALLET_NETWORKS.get(
-    chainId
-  )}`
-
-  try {
-    address = toChecksumAddress(address.trim())
-  } catch (err) {
-    return ''
-  }
-
-  if (address === EMPTY_ADDRESS) {
-    return `${baseUrl}/info/logo.png`
-  }
-
-  return `${baseUrl}/assets/${address}/logo.png`
 }
